revise code nomor 1

+++ Lab01/Q1.cpp
@@ -22,7 +22,7 @@ int main()
   multiplication = a * b;
 
   //division
-  division = a / b;
+  division = (double)(a / b);
 
   //remainder
   remainder = a % b;
